+++ ts2net/__init__.py
@@ -22,6 +22,11 @@ from .api import HVG, NVG, RecurrenceNetwork, TransitionNetwork, build_network
 from .core import graph_summary
 from .exceptions import NotBuiltError, ValidationError, Ts2NetError
 from .protocols import NetworkBuilder
+# Redundant aliases mark these as intentional re-exports (PEP 484).
+from .state import StateSpace as StateSpace
+from .state import reconstruct as reconstruct
+from .ccm import ccm as ccm
+from .ccm import ccm_test as ccm_test
 
 __version__ = "0.9.0"
 
@@ -79,6 +84,7 @@ except ImportError:
 # Windowed graphs API
 from .api_windows import build_windows
 __all__.append('build_windows')
+__all__.extend(['reconstruct', 'StateSpace', 'ccm', 'ccm_test'])
 
 # Multiscale graphs API
 from .multiscale import MultiscaleGraphs, coarse_grain